Make vision heading a double instead of an int & fix heading orientation
Also add some debugging

+++ src/org/usfirst/frc/team1076/robot/commands/TurnWithVision.java
@@ -28,7 +28,7 @@ public class TurnWithVision extends Command {
     @Override
     public boolean execute() {
         receiver.receive();
-        int heading = receiver.getData().getHeading();
+        double heading = receiver.getData().getHeading();
         int errorCount = receiver.getData().getErrorCount();
         /* The heading essentially acts as the direction to turn.
          * Negative = left, positive = right.
